@@ -2,6 +2,7 @@
   "name": "journal-app",
   "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
@@ -12,23 +13,26 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "@supabase/supabase-js": "^2.39.0",
+    "@supabase/supabase-js": "^2.39.1",
     "bootstrap": "^5.3.2",
     "pinia": "^2.1.6",
-    "sass": "^1.69.4",
-    "vue": "^3.3.4",
-    "vue-router": "^4.2.4"
+    "sass": "^1.69.5",
+    "vee-validate": "^4.12.4",
+    "vue": "^3.3.13",
+    "vue-router": "^4.2.4",
+    "vue3-toastify": "^0.2.1",
+    "yup": "^1.3.3"
   },
   "devDependencies": {
-    "@commitlint/cli": "^18.0.0",
-    "@commitlint/config-conventional": "^18.0.0",
-    "@vitejs/plugin-vue": "^4.3.4",
-    "@vue/test-utils": "^2.4.1",
-    "eslint": "^8.49.0",
+    "@commitlint/cli": "^18.4.3",
+    "@commitlint/config-conventional": "^18.4.3",
+    "@vitejs/plugin-vue": "^5.0.2",
+    "@vue/test-utils": "^2.4.3",
+    "eslint": "^8.56.0",
     "eslint-plugin-vue": "^9.17.0",
     "husky": "^8.0.0",
     "jsdom": "^22.1.0",
-    "vite": "^4.4.9",
+    "vite": "^5.0.10",
     "vitest": "^0.34.4"
   }
 }
